Update Android app for Mapbox 11.18, OTP 2.7.0, and GCP deployment
- Upgrade Mapbox SDK to 11.18.2 with updated Compose API (trailing lambdas, Color objects)
- Fix OTP GraphQL query for 2.7.0 schema (location wrapper, transit modes structure)
- Fix JSON deserialization: OTP returns leg duration as Double, not Int
- Point OTP base URL to GCP VM (35.197.42.97) and allow cleartext traffic
- Upgrade Gradle wrapper to 9.3.1
- Add launcher icon resources and Mapbox access token
- Add OTP docker-compose config

+++ app/src/main/java/com/skilift/app/data/remote/dto/OtpResponse.kt
@@ -42,7 +42,7 @@ data class LegDto(
     val mode: String,
     val startTime: Long? = null,
     val endTime: Long? = null,
-    val duration: Int? = null,
+    val duration: Double? = null,
     val distance: Double? = null,
     val from: PlaceDto,
     val to: PlaceDto,
